Consolidate all graphing-related files, except GraphActivity
