+++ app/src/org/commcare/android/view/EntityDetailView.java
@@ -18,12 +18,13 @@ import android.widget.TextView;
 
 import org.commcare.android.javarosa.AndroidLogger;
 import org.commcare.android.models.Entity;
-import org.commcare.android.models.graph.GraphData;
 import org.commcare.android.util.DetailCalloutListener;
 import org.commcare.android.util.FileUtil;
 import org.commcare.android.util.InvalidStateException;
 import org.commcare.android.util.MediaUtil;
 import org.commcare.dalvik.R;
+import org.commcare.graph.model.GraphData;
+import org.commcare.graph.view.GraphView;
 import org.commcare.suite.model.CalloutData;
 import org.commcare.suite.model.Detail;
 import org.javarosa.core.reference.InvalidReferenceException;
