+++ app/src/org/commcare/android/view/EntityView.java
@@ -18,13 +18,14 @@ import android.widget.TextView;
 
 import org.commcare.android.models.AsyncEntity;
 import org.commcare.android.models.Entity;
-import org.commcare.android.models.graph.GraphData;
 import org.commcare.android.tasks.ExceptionReporting;
 import org.commcare.android.util.AndroidUtil;
 import org.commcare.android.util.InvalidStateException;
 import org.commcare.android.util.MediaUtil;
 import org.commcare.android.util.StringUtils;
 import org.commcare.dalvik.R;
+import org.commcare.graph.model.GraphData;
+import org.commcare.graph.view.GraphView;
 import org.commcare.suite.model.Detail;
 import org.javarosa.core.services.Logger;
 import org.javarosa.core.services.locale.Localization;
