+++ app/src/org/commcare/graph/model/AnnotationData.java
@@ -1,4 +1,4 @@
-package org.commcare.android.models.graph;
+package org.commcare.graph.model;
 
 /**
  * Data for an annotation, which is text drawn at a specified x, y coordinate on a graph.
