+++ app/src/org/commcare/graph/model/BubblePointData.java
@@ -1,4 +1,4 @@
-package org.commcare.android.models.graph;
+package org.commcare.graph.model;
 
 /**
  * Representation of a point on a bubble chart, which has an x, y position and an additional value for the bubble's radius.
