+++ app/src/org/commcare/graph/model/GraphData.java
@@ -1,4 +1,4 @@
-package org.commcare.android.models.graph;
+package org.commcare.graph.model;
 
 import java.util.Hashtable;
 import java.util.Vector;
@@ -46,7 +46,7 @@ public class GraphData implements ConfigurableData {
 
     /*
      * (non-Javadoc)
-     * @see org.commcare.android.models.graph.ConfigurableData#setConfiguration(java.lang.String, java.lang.String)
+     * @see org.commcare.graph.model.ConfigurableData#setConfiguration(java.lang.String, java.lang.String)
      */
     public void setConfiguration(String key, String value) {
         mConfiguration.put(key, value);
@@ -54,7 +54,7 @@ public class GraphData implements ConfigurableData {
 
     /*
      * (non-Javadoc)
-     * @see org.commcare.android.models.graph.ConfigurableData#getConfiguration(java.lang.String)
+     * @see org.commcare.graph.model.ConfigurableData#getConfiguration(java.lang.String)
      */
     public String getConfiguration(String key) {
         return mConfiguration.get(key);
@@ -62,7 +62,7 @@ public class GraphData implements ConfigurableData {
 
     /*
      * (non-Javadoc)
-     * @see org.commcare.android.models.graph.ConfigurableData#getConfiguration(java.lang.String, java.lang.String)
+     * @see org.commcare.graph.model.ConfigurableData#getConfiguration(java.lang.String, java.lang.String)
      */
     public String getConfiguration(String key, String defaultValue) {
         String value = getConfiguration(key);
