+++ app/src/org/commcare/graph/view/c3/LegendConfiguration.java
@@ -1,6 +1,6 @@
-package org.commcare.android.view.c3;
+package org.commcare.graph.view.c3;
 
-import org.commcare.android.models.graph.GraphData;
+import org.commcare.graph.model.GraphData;
 import org.json.JSONException;
 
 /**
